+++ crates/vyxen/src/prelude.rs
@@ -6,3 +6,4 @@ pub use vyxen_math::*;
 pub use vyxen_physics2d::*;
 pub use vyxen_renderer::*;
 pub use vyxen_resource::*;
+pub use vyxen_ui::*;